app/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
aiohttp
tenacity
redis
python-multipart
asyncpraw